Reorganizing main codebase.
