+++ src/commands/mod.rs
@@ -1,0 +1,4 @@
+pub mod log_commands;
+pub mod media_commands;
+pub mod server_commands;
+pub mod user_commands;
